rave: Add bootstrap system.
The bootstrap system allows rave to be, as the name implies, bootstrapped
using various means. Currently only filesystem-based bootstrapping is supported,
but that may change in the future.

Bootstrapping is done in two stages: the global engine bootstrap,
and the game-specific bootstrap. The engine is bootstrapped using
rave.bootstrap.bootstrap_engine(), the game using rave.bootstrap.bootstrap_game().
Both functions take a bootstrapper name, if none is given an attempt will be made
to auto-detect the bootstrapper.

Bootstrappers should be placed in rave/bootstrappers/ and should implement the
following API:
- bootstrap_modules(): bootstrap essential engine modules.
- bootstrap_filesystem(): bootstrap engine file system and mount rave.bootstrap.ENGINE_MOUNT,
    rave.bootstrap.MODULE_MOUNT and rave.bootstrap.COMMON_MOUNT on the file system.
- bootstrap_game_filesystem(): bootstrap game file system and mount rave.bootstrap.GAME_MOUNT.